Add GOWORK env variable when go vet is run
The cinder-operator patch [1] introduces the 'go.work' file within the
repository with the goal of automating the way we deal with multi module
dependencies.
However, the side effect of having a versioned 'go.work' in the repo is
the failure of this script because GOWORK is automatically set to "on".
This patch fixes the CI run on the cinder-operator PR [1], introducing
a GOWORK variable (which is set to off by default) to make sure we can
successfully run this check regardless of the Makefile approach.

[1] https://github.com/openstack-k8s-operators/cinder-operator/pull/79

Signed-off-by: Francesco Pantano <fpantano@redhat.com>