[renovate]Allow running make tidy
The service operator repos all have `make tidy` targets now to run go
mod tidy on both mod files in the repo. So to be able to use that in
renovate this patch adds that command to the allow list.

Also lib-common is now changed to have make tidy instead of make go
tidy. As that was the last user of make gotidy, that command is removed
from the allow list now.