Limit auto rebase of renovate PRs
Until now renovate rebased its open PRs every hour if there was a new
version from the dependency being bump or if there was a new change on
the base branch of the PR. This is fairly wasteful for CI resources.

This change instructs renovate to only auto rebase the PRs during our
weekend schedule when new PRs are also considered.

The user can still request rebasing a PR via the checkbox in the PR
description.