Add force-bump-pull-request workflow
This workflow calls force-bump and generates a pull request
which should contain all the necissary updates to manifests.
For openstack-operator bindata is also updated.

If no changes are detected no PR should get generated

Includes an additional force-bump-branches workflow which
can be called via a schedule job in the operator repos
to schedule PRs for all stable braches (main, fr2, etc)
to be automatically generated on weekends.

Jira: OSPRH-15737