[force-bump] make sure modules are synced
In bump PRs it happened that the dependencies are not in sync
in ./go.mod and ./api/go.mod. This makes the pre-commit job to
fail.
This change make sure we init the go.work before running force-bump
and call a `go work sync` afterwards. This should prevent those
diff in ./go.mod and ./api/go.mod.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>