Convert latesttag to lowercase
The redhat-actions/push-to-registry@v2 action requires images with
lowercase tags. If a tag is passed that is not lowercase, it converts it
to lowercase:

---
2025-12-12T11:43:23.0373178Z ##[warning]Reference to image and/or tag
must be lowercase. Reference has been converted to be compliant with
standard.
---

However, if the image wasn't actually tagged with an all lowercase tag
in the previous step, then the action fails to find the image:

---
2025-12-12T11:43:23.3474663Z ##[error]❌ All tags were not found in
either Podman image storage, or Docker image storage. Tag
"openstack-operator-index:osprh-17457-openstackdataplaneservice_overrides_labels-latest"
not found in Podman image storage, and tags
"openstack-operator-index:osprh-17457-openstackdataplaneservice_overrides_labels-latest,
openstack-operator-index:68acc12e8b2b5213a958c30981d544deefe58cd0" not
found in Docker image storage.
---

This change converts the "latesttag" env var to lowercase before tagging
the image so that it can be found by the subsequent action.

Jira: OSPRH-23119
Signed-off-by: James Slagle <jslagle@redhat.com>